better time management while playing games
Idea comming from [HGM's post on TalkChess](https://talkchess.com/forum3/viewtopic.php?f=7&t=76463#p881078).

As HGM explained, we should not abort a search iteration !
I found a way to predict if the next iteration would be aborted by timeout using an exponential regression.

I did not apply his idea of continue a search that should be aborted if the score change dramatically. Maybe another day !

In the following recap of a match of 100 games in 40/10+0.1, the new version is called Reglisse-TM (because of it better _Time Management_ during the test) and scored with a positive Elo gain (+74.1 +/- 67.9). As we can see, with this update, the engine almost never lose on time.

```
Score of Reglisse-JS vs Reglisse-TM: 36 - 57 - 7  [0.395] 100

Elo difference: +74.1 +/- 67.9, LOS: 1.5 %, DrawRatio: 7.0 %

Reglisse-TM
    Draw by 3-fold repetition : 1
    Draw by fifty moves rule : 3
    Draw by insufficient mating material : 3
    Loss by mate : 36
    Win by timeout : 6
    Win by mate : 51
```

What I plan to do, if I'm able find the time to do it :

- Remove some tag changes in make/unmake null-moves that are useless. I already tested it quickly, by I need to test it deeper to make sure this would not create any bugs. Should improve time performances.

- Use piece-lists in order to track where are the pieces of each color, to avoid looping over all 64 square while we chan loop overt 32 (evaluation) to 16 (move-generation) squares maximum. Should also improve time performances, maybe drastically.